+++ lua/chatforge/init.lua
@@ -13,7 +13,7 @@
  
 local M = {}
 
-M.version = "0.2.0"
+M.version = "0.2.1"
  
 function M.setup(opts)
   local config  = require("chatforge.config")